tests: use ReTest package instead of Test
